feat: adding proxy service for better concern separation
