+++ services/__init__.py
@@ -1,0 +1,1 @@
+"""Service layer for ollama-mcp-bridge."""
